# Resolved conflicts:
#	TapBand_SRC/Assets/Editor/Scripts/GameDataLoader/RawGameDataLoader.cs
#	TapBand_SRC/Assets/Script/Controllers/CurrencyController.cs
#	TapBand_SRC/Assets/Script/Controllers/MerchController.cs
#	TapBand_SRC/Assets/Script/Model/GameData.cs
#	TapBand_SRC/Assets/Script/Model/GameState.cs
#	TapBand_SRC/Assets/Script/Model/Sheets/ConcertData.cs
#	TapBand_SRC/Assets/Script/Model/Sheets/MerchData.cs
#	TapBand_SRC/Assets/Script/Model/Sheets/MerchSlotData.cs
#	TapBand_SRC/Assets/Script/Model/Sheets/MerchSlotData.cs.meta
#	TapBand_SRC/Assets/Script/Model/State/CurrencyState.cs
#	TapBand_SRC/Assets/Script/Model/State/MerchState.cs
#	TapBand_SRC/Assets/Script/Views/MerchUI.cs

+++ TapBand_SRC/Assets/Script/Model/Sheets/MerchSlotData.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 public class MerchSlotData
 {
     public int id;
-    public int coinCost;
+    public double coinCost;
     public int tokenCost;
 }
+
